chore: Release v0.18.0 with CUDA support (Refs #77)
- Bump version 0.17.0 → 0.18.0
- Update trueno dependency 0.8.1 → 0.8.3
- Add cuda feature using trueno/cuda-monitor
- Add LoadConfig::cuda() and LoadConfig::gpu() presets
- Add Backend::Cuda variant with helper methods
- Add comprehensive tests for CUDA/GPU backends
- Coverage improvements across multiple modules
- Update codecov.yml to exclude local paths

+++ src/text/summarize.rs
@@ -310,12 +310,12 @@ impl TextSummarizer {
         for i in 0..n {
             for j in 0..n {
                 if i != j {
-                    let intersection: usize = tokenized[i].intersection(&tokenized[j]).count();
+                    let intersection: f64 = tokenized[i].intersection(&tokenized[j]).count() as f64;
 
                     let union_size = tokenized[i].len() + tokenized[j].len();
 
                     if union_size > 0 {
-                        similarity[i][j] = (2.0 * intersection as f64) / union_size as f64;
+                        similarity[i][j] = (2.0 * intersection) / union_size as f64;
                     }
                 }
             }
